Rewrite integration test to insert test message data by reading from JSON files in resources folder
